Rework how we build snapshots
When we build a snapshot we want to build it with the clang that
ships with Fedora. What happened before was that we've build clang with
the latest version of clang that we've built ourselves. This is of
course unfortunate.

This is an approach to have three copr projects:

0. `llvm-snapshots-incubator-<YESTERDAY>` (will be kept for 7 days)
1. `llvm-snapshots-incubator-<TODAY>` (will be kept for 7 days)
2. `llvm-snapshots` (will be overwritten each day with `llvm-snapshots-incubator-<TODAY>`)

The `llvm-snapshots` project is *forked* from the
`llvm-snapshot-incubator-<YESTERDAY>` over and over again every day. This copies
over successful builds from the `llvm-snapshots-incubator-<YESTERDAY>` project.

The advantages are:

 * Every day we build in a fresh project without any packages from previous builds.
 * We can look back 7 days and see projects.
 * We don't have to tell users to change how they get the snapshots.

Downsides:

 * Collecting statistics need to be adjusted.
 * Lot more juggling with projects.

I hope this makes sense and looks right.

NOTICE: For now I'll do this under my ownername (`kkleine` and not under `@fedora-llvm-team`)