Ensure all combinations of package+chroot exist
This ensures that all combinations of packages and chroots that we care
about, do exist in in the monitor of a project.

Before we just checked for something other than `succeeded` builds but
it could be that in a project there's no build at all or a build is
missing. Then we don't want to wave a build through.