Simplify injection of JS and full-height diagrams
I remembered that at one point in time we had diagrams always expand to
the full height of the browser and in my last change (#58) I increased
the height from 450px to 800px to make it better and more like before.
Then I realized that the method I've used before to generate the
diagrams using `fig.write_html()` was less complicated but had
full-height diagrams. By inspecting the methods parameters I found that
it had a way to inject JavaScript which is exactly what we needed in
order to make the points on a line clickable. I had to remove the
`<script></script>` tag around the JS snippet and replace the `<div>`'s
ID with a placeholder `{plot_id}` which will be replaced by plotly.