check-snapshots workflow: use matrix strategy
I found myself duplicating things again for the upcoming bootstrap
builds and figured it is best to employ a matrix strategy for
checking for snapshots. That is easily extensible. Just consider moving
from big-merge to bootstrap and then to PGO. All will benefit from this
change. I'll do the same for the rest of the jobs soon.

Additionally I've added strategy labels (`strategy/standalone` and
`strategy/big-merge`) labels to issues that report a broken snapshot
build. This allows us to better filter for already existing issues.