error context information and more
A `error/test' cause was added.

With this change we collect an actual excerpt from the build log for
each error that we can identify. The snippet shows the surroundings of
each error marker (e.g. a regex pattern). By definition we don't have a
way to grep for an unknown error. Therefore we use the tail of the build
log as the error context snippet.

Based on the request from @nikic, for the "unknown" build failure, we
create a comment in the issue on github that identifies the broken
snapshot build. The comment contains a link to the build log (NOTE: it
might expire on Copr) and a materialized excerpt from the build log. In
theory we could do this for every error that we can or cannot identify
but it was only requested for the "unknown" cause so far.

Since we run the checks for snapshots periodically, I've
added a mechanism to only post such a comment once for each
package+chroot+cause combination. This is done by simply adding an HTML
comment to the beginning of the comment that I can search for.