tarball-generation: find good commit again (#221)
We go over the last 30 commits and check if there's one that passes all of our required tests and that has a successful combined status. From all the successful commits we'll choose the one that has the most other checks passing without explicitly looking at the kind of checks.

The idea is to not build commits in our snapshots that are known to be broken. By limiting the number of commits to 30 we should stay within the boundaries of the day and not go too far back.

The output of `get-good-commit.py` is now split in two: `stdout` and `stderr` for us to get good logs on `stderr` and also get the final output commit SHA on `stdout`.