Add and use prepare-python action
This action checks out the repository and configures the python
environment and makes use of a cache. This way we can avoid code
duplication and split our workflows into more than one job more easily.
See the portion that deletes the old snapshots for example. That is done
in parallel to generating the new snapshots now.