Downgrade testing-farm to 0.0.16
This should allow us to get artifact URLs pointing to testing-farm results on the redhat ranch.

In 0.0.18 or before, testing-farm introduced more detailed output in the
`testing-farm watch` command which required to make calls that at least
the redhat ranch were only possible in the Red Hat VPN which is not
possible on github.