Add some automation for doing periodic mass rebuilds of Fedora packages (#864)
These mass rebuilds will help test the snapshot builds and catch
regressions in real-world projects.  The GitHub workflows are
currently configured to start a new mass rebuild once per month
and then when the rebuild is complete, it will create any issue
reporting any regressions since the last build.

The goal is to keep the rebuilder as simple as possible, so it will
only rebuild packages that successfully built in the last rebuild, and
it makes no attempt to work around packages that hard code gcc as their
compiler.  As a result, some of the 'passing' builds may actually use
gcc instead of clang, but this is an acceptable trade-off to keep
the rebuild process simple.

Co-authored-by: Konrad Kleine <konrad.kleine@posteo.de>