Remove project/... label functionalities
We will no longer add `project/...` labels to our issues because we're
only building the one big `llvm` project for a while now. We could
invest in better classification of LLVM sub-projects as described in
[this issue](#579) but that's a different story. For now, let's remove
the `project/...` label facility.

NOTE: In places where the testing code was using `project/clang` for
example, I've picked another label that doesn't interfere with the test.

Fixes: #1103
See: #579