Limit the number of fedora-s390x chroots
COPR restricts the number of s390x builds each team can start, queueing
builds after this limit is reached. This is causing daily builds to take
too long to complete.

In order to minimize this, we decided to limit the number of
fedora-s390x chroots to 2: rawhide and the latest branched version.

Add an exclude mechanism to get_chroots() that allows to apply this
limit.