Use chroots from augmented config in snapshot_manager
In a previous commit we ensured that a `config.Config` object not only
knows about the `chroot_pattern` but also can carry the resolved
`chroots` (#1145) that match this pattern. The methods to filter and
sanitize chroots was added in (#1146).

In this change we make use of the `config.Config.chroots` and remove the
need to fetch and filter.

ghstack-source-id: 9e103acc7f47298dfb1c638018cedc147a61c170
Pull Request resolved: https://github.com/fedora-llvm-team/llvm-snapshots/pull/1148