Delete no longer needed files
We no longer need these files:

* `.github/actions/generate-matrix/action.yml`
* `scripts/functions-pgo.sh`
* `scripts/functions.sh`

This is the last step in moving functions from `scripts/functions.sh` to
python and consolidate the configuration in python.

As a bonus, the `generate-matrix` action is removed. Actions don't have
access to secrets and therefore most of the action consisted of inputs
that map to the secrets. We can now do the same with a simply call to
`python3 snapshot_manager/main.py github-matrix`.

ghstack-source-id: 7c7f9af50dfa3e0afc88c34e4e4af6e40d93410f
Pull Request resolved: https://github.com/fedora-llvm-team/llvm-snapshots/pull/1151