Attempt to mitigate lsb_release error (#1163)
We saw this error before:

```
Error: Unable to locate executable file: lsb_release. Please verify either the file path exists or the file can be found within a directory specified by the PATH environment variable. Also check the file mode to verify the file is executable.
```

The root cause is that the `prepare-python` action is tailored to run on
`ubuntu` containers. As a short term fix for this I've implemented the
steps manually in the `fedora-copr-build.yml` workflow which runs on a
`fedora:41` container.