Run fedora-copr-build.yml in ubuntu container (#1165)
There's no need to install copr-cli using dnf when we have it in our
list of python requirements. The only difference is that we need to call
it with `copr-cli` instead of just `copr`. As a matter of fact `/usr/bin/copr`
on Fedora Linux is just a link to `/usr/bin/copr-cli`.

This should allow us mitigate problems with differing container setups
and their usage with `actions/setup-python`.

This was the last workflow to use a fedora container.