Add mechanism to compare compile-time perf (#672)
This change adds a directory `script/perf` that contains files to run a
compile-time performance test for snapshots (called "big-merge"), PGOed
snapshots setup and the system LLVM.

To run the performance test for the current day, run the following steps:

```console
$ cd scripts/perf
$ make
```

You'll be prompted with a markdown output that you can copy and paste into a
github issue.

The output is located between `<!--BEGIN REPORT-->` and `<!--END REPORT-->`
markers.