Enable same chroots for PGO as for big-merge (#1190)
We will build PGO for non x86 for archs and for fedora >= 41 and RHEL >=
9. But we still want to build for the same chroots as big-merge because
we need to make sure we can build for those archictures with the PGO
code in the llvm.spec not breaking the normal build.

The expected compile-time performance improvements for x86, fedora < 41
and RHEL8 chroots has to be identical to the current snapshots.