Add test plan for compile-time perf. comparison
This test plan compares three versions of LLVM and their compile time:

  1. The PGOed LLVM snapshot living in the daily copr repo:
     `llvm-snapshots-pgo-$YYYYMMDD`
  2. The regular LLVM snapshot also living in another daily copr repo:
     `llvm-snapshots-big-merge-$YYYYMMDD`
  3. The system LLVM living in the regular fedora repository.

NOTE: The comparison happens on the same machine for each chroot to be fair.
We store all comparison results in the test plans artifact directory (see
`TMT_PLAN_DATA` in
https://tmt.readthedocs.io/en/stable/overview.html#step-variables). Expect
these files for each passing plan:

  1. pgo.json
  2. snapshot.json
  3. system.json
  4. pgo_vs_snapshot.compile_time.txt
  5. pgo_vs_system.compile_time.txt
  6. snapshot_vs_system.compile_time.txt
  7. result.csv

In order to run this test plan on testing-farm you need to make a request like
so:

```console
$ testing-farm request \
  --compose Fedora-Rawhide \
  --git-url https://github.com/kwk/llvm-snapshots \
  --git-ref compile-time-perf \
  --arch x86_64 \
  --plan /tests/compare-compile-time \
  --context distro=fedora-rawhide \
  --context arch=x86_64 \
  --environment COPR_CHROOT=fedora-rawhide-x86_64 \
  --environment YYYYMMDD=20250317
```

What's really necessary to pass is the `YYYYMMDD` environment variable in
order to pick the right copr repos and the `COPR_CHROOT` environment variable
to ensure that copr repo enablement works as expected.

NOTE: This plan was designed to be flexible enough to re-use it for testing
      i.e. BOLT performance improvements, once we have a copr repo which
      enables it.

NOTE: In order for the already existing scripts in `scripts/perf` I've
      refactored the bash functions from `scripts/perf/entrypoint.sh` to
      be re-usable and put them in `scripts/perf/lib.sh`. So now, the
      local and `podman`-based performance comparison runs using the
      same functions as we use for the testing plan.

Outlook
-------

In a future change I'm going to implement the logic for finding two
pairs of copr repositories that we want to compare against and then make
a function call to create the testing-farm request.