[workflow] performance diagrams
I've renamed the `update-build-time-diagrams.yml` to
`update-diagrams.yml` because we want to re-use the logic for adding
a CSV file (`perf-results.csv`) to the `gh-pages` branch.

In this workflow we collect the results of the performance comparison
for "yesterday".

The new `collect-perf-results` command of the `main.py` will be
introduced in a later PR.

The new `create-performance-diagrams.py` file will be introduced in a
later PR.

commit-id:36304537