Refactor and add make_snapshot_gating_request
There was more and more testing-farm related code in
`testing_farm_util.py` that it made sense to split it up into multiple
modules.

Some functionality that is shared between modules is located in
`tfutil.py`.

The `make_compare_compile_time_request` is probably the most important
change in this PR because that is the one that will kick off the
`compare-compile-time.fmf` plan to run. Tests for this function have
already been added in a previous PR entitled "[tests] Entrypoint fns for
run + collect"

I didn't mention this before but PRs before this one actually made
use of this refactoring. For example the "perf-diagrams" command added
before is adding an extra column `testing_farm_artifacts_url` to its
internal `pandas.DataFrame` that required information from testing-farm.

commit-id:b1f3828d