mass-rebuild-reporter: Rebuild failed packages in COPR
Before we report a failure as a regression, rebuild it in COPR to make
sure that it is a true failure.  The benefit of doing this in COPR
instead of locally, is that if the build passes then it will become
part of the baseline packages for the next rebuild.  Otherwise, if
the latest build in COPR is still a failure, then the package won't
be rebuilt during the next mass rebuild.