Build daily assets for llvm-test-suite
There's a new github workflow that will generate these assets for the
`llvm-test-suite-version-sync` release [1].

The new `sync-on-llvm-test-suite-version.yml` workflow generates these
files daily:

 1. `llvm-test-suite-release-<YYYYMMDD>.txt`
 2. `llvm-test-suite-git-revision-<YYYYMMDD>.txt`
 3. `llvm-test-suite-version-<YYYYMMDD>.txt`

The first file contains the LLVM upstream release for today
(e.g. `22.0.0`). The second file contains the upstream's
`llvm-test-suite` Git hash (e.g. `f302e7778e4a0bb58b40a3b0cabde719e33fb55e`).
The last file is somewhat special because it can be used as a
`version.spec.inc` in the `llvm-test-suite.spec` right away. It looks
like this:

```rpmspec
%global llvm_test_suite_version 22.0.0
%global llvm_test_suite_date 20251109
%global llvm_test_suite_git_revision f302e7778e4a0bb58b40a3b0cabde719e33fb55e
%global llvm_test_suite_git_revision_short f302e7778e4a0b
%{error: No need to build today. llvm-test-suite's git revision (f302e7778e4a0b) and the LLVM version (22.0.0) were the same yesterday.}
```

We write this file every day but we add the last line at the very bottom
only if we think that a re-build of the same source is not needed. This
gives the SRPM generation a chance to abort a build.

See [this llvm-test-suite PR](https://src.fedoraproject.org/rpms/llvm-test-suite/pull-request/58) for how the assets will be used.

[1]: https://github.com/fedora-llvm-team/llvm-snapshots/releases/tag/llvm-test-suite-version-sync