Create daily llvm-test-suite builds (#1759)
* Create daily llvm-test-suite builds

With a few tweaks I was able to re-use most of the `fedora-copr-build`
workflow for the `llvm-test-suite`.

We don't start a build of llvm-test-suite in the first place if
we can tell that it is not needed.

The PR that creates the llvm-test-suite assets MUST go first: #1766.
At best with a couple of days before this one gets merged.

* Only skip build steps, not promotion for llvm-test-suite

Suppose yesterday was a successful build but nothing has changed today.
Then we don't want to build llvm-test-suite today. BUT we still want to
promote the build from yesterday and that was skipped before this
commit.