Add "run_check_snapshots_workflow" to config
Then use this to check if we want to run the `check-snapshots.yml`
workflow. For `llvm-test-suite` I don't want to create daily issues now.
The main reason is that we don't have daily builds. We only build
`llvm-test-suite` when things have changed.

The current system is designed to query for the daily copr repo and that
would fail when we don't have daily builds.